created helper.js file to store function that needed to be called inside the index.html file
right now these are copying the join and watch link to clipboard,
through the rely on the existence of the join and watch <a
href=> elements.

Could be made more robust by generating the join and watch links for
the game and storing them someplace.
Or also, removing the open in new tab buttons and instead changing
the lines in recieveMessages to update the text field of the
element, which then calls a function to copy to clipboard.

Now will test links on Github pages